// packages/browser-extension/src/store/extension/mutation-types.ts
export enum MutationTypes {
  ADD_PENDING = 'ADD_PENDING',
  REMOVE_REQUEST = 'REMOVE_REQUEST',
  SET_WALLET = 'SET_WALLET',
  SET_LAST_ACCOUNT = 'SET_LAST_ACCOUNT',
  SET_NEW_ACCOUNT = 'SET_NEW_ACCOUNT',
  SET_KEY_HASHES = 'SET_KEY_HASHES',
  SET_WHITELISTED_WEBSITES = 'SET_WHITELISTED_WEBSITES',
}